Added esxi driver

+++ dcmgr/lib/dcmgr/rpc/hva_handler.rb
@@ -69,7 +69,7 @@ module Dcmgr
         end
         
         # cleanup vm data folder
-        FileUtils.rm_r(File.expand_path("#{@inst_id}", @node.manifest.config.vm_data_dir))
+        FileUtils.rm_r(File.expand_path("#{@inst_id}", @node.manifest.config.vm_data_dir)) unless @hv.is_a?(Dcmgr::Drivers::ESXi)
       end
 
       def update_instance_state(opts, ev)
@@ -241,62 +241,48 @@ module Dcmgr
         @inst = rpc.request('hva-collector', 'get_instance',  @inst_id)
         raise "Invalid instance state: #{@inst[:state]}" unless %w(pending failingover).member?(@inst[:state].to_s)
 
-        # select hypervisor :kvm, :lxc
+        # select hypervisor :kvm, :lxc, :esxi
         select_hypervisor
 
         # create hva context
         @hva_ctx = HvaContext.new(self)
 
+        Drivers::LocalStore.select_local_store(@hv.class.to_s.downcase.split('::').last).deploy_image(@inst,@hva_ctx)
+
         rpc.request('hva-collector', 'update_instance', @inst_id, {:state=>:starting})
-        # setup vm data folder
-        inst_data_dir = @hva_ctx.inst_data_dir
-        FileUtils.mkdir(inst_data_dir) unless File.exists?(inst_data_dir)
-        # copy image file
-        img_src = @inst[:image][:source]
-        @os_devpath = File.expand_path("#{@inst[:uuid]}", inst_data_dir)
-
-        # vmimage cache
-        vmimg_cache_dir = File.expand_path("_base", @node.manifest.config.vm_data_dir)
-        FileUtils.mkdir_p(vmimg_cache_dir) unless File.exists?(vmimg_cache_dir)
-        vmimg_basename = File.basename(img_src[:uri])
-        vmimg_cache_path = File.expand_path(vmimg_basename, vmimg_cache_dir)
-
-        logger.debug("preparing #{@os_devpath}")
-
-        # vmimg cached?
-        unless File.exists?(vmimg_cache_path)
-          logger.debug("copying #{img_src[:uri]} to #{vmimg_cache_path}")
-          pararell_curl("#{img_src[:uri]}", "#{vmimg_cache_path}")
-        else
-          md5sum = sh("md5sum #{vmimg_cache_path}")
-          if md5sum[:stdout].split(' ')[0] == @inst[:image][:md5sum]
-            logger.debug("verified vm cache image: #{vmimg_cache_path}")
-          else
-            logger.debug("not verified vm cache image: #{vmimg_cache_path}")
-            sh("rm -f %s", [vmimg_cache_path])
-            tmp_id = Isono::Util::gen_id
-            logger.debug("copying #{img_src[:uri]} to #{vmimg_cache_path}")
-            pararell_curl("#{img_src[:uri]}", "#{vmimg_cache_path}.#{tmp_id}")
-
-            sh("mv #{vmimg_cache_path}.#{tmp_id} #{vmimg_cache_path}")
-            logger.debug("vmimage cache deployed on #{vmimg_cache_path}")
-          end
-        end
-
-        ####
-        logger.debug("copying #{vmimg_cache_path} to #{@os_devpath}")
-        case vmimg_cache_path
-        when /\.gz$/
-          sh("zcat %s | cp --sparse=always /dev/stdin %s",[vmimg_cache_path, @os_devpath])
-        else
-          sh("cp -p --sparse=always %s %s",[vmimg_cache_path, @os_devpath])
-        end
 
         sleep 1
 
+        ###### Setup metadata for ESXi ######
+        inst_data_dir = @hva_ctx.inst_data_dir
+        FileUtils.mkdir(inst_data_dir) unless File.exists?(inst_data_dir)
         setup_metadata_drive
         
-        check_interface
+        #@hv.setup_metadata_iso if @hv.is_a?(Dcmgr::Drivers::ESXi)
+        if @hv.is_a?(Dcmgr::Drivers::ESXi)
+          begin
+          sh("genisoimage -R -o #{inst_data_dir}/metadata.iso #{@hva_ctx.metadata_img_path}")
+          #sh("mkdir #{inst_data_dir}/#{@hva_ctx.metadata_img_path}")
+          #sh("")
+          esxi_options = {
+            :host => @hva_ctx.node.manifest.config.esxi_ipaddress,
+            :user => @hva_ctx.node.manifest.config.esxi_username,
+            :password => @hva_ctx.node.manifest.config.esxi_password,
+            :insecure => true,
+            :datastore => @hva_ctx.node.manifest.config.esxi_datastore,
+            :datacenter => @hva_ctx.node.manifest.config.esxi_datacenter,
+          }
+          
+          sh("curl -s -u #{esxi_options[:user]}:#{esxi_options[:password]} -k -T #{inst_data_dir}/metadata.iso https://#{esxi_options[:host]}/folder/#{@hva_ctx.inst[:uuid]}/metadata.iso?dsName=#{esxi_options[:datastore]}")
+          ensure
+            sh("rm -rf #{inst_data_dir}")
+          end
+        end
+        ###### Finished setting up metadata drive. TODO: REFACTOR THIS! ######
+        
+        
+        
+        #check_interface
         @hv.run_instance(@hva_ctx)
         update_instance_state({:state=>:running}, 'hva/instance_started')
       }, proc {
